+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # MeshAdminUI Changelog
 
+## Version 0.8.2 (tbd)
+* Add multi binary upload dialog. Users can now upload multiple files at once by clicking the button next to the create node button.
+
 ## Version 0.8.1 (2017-11-29)
 * Fix trying to publish after failed save. This has been hiding the error message of the failed save.
 * Fix dropdown positioning in IE